The bones
The first iteration of it. It didn't take long to make but has taught me some very important things in its creation. Local MongoDB authentication, storing and retrieving posts while ordering them by date and allowing a search option. A hidden option that works, hiding the post from anyone who isn't an admin or the owner. Only the owner or an admin can delete or edit the posts.
Written in javascript, using EJS, Passport, Mongoose, and many other plugins. A very small amount of Bootstrap is present but barely.
(PS: Not the best at designing, so if you decide to use this please give it some love because it looks egregious at present date.)